requirements.txt
crewai
crewai_tools
python-dotenv
langchain-huggingface
